fix: analysisRepository 로직 수정

+++ backend/src/main/java/com/buildsync/dto/analysis/MonthlyMaterialCostProjection.java
@@ -4,7 +4,7 @@ public interface MonthlyMaterialCostProjection {
 
 	String getMonth();
 	
-	Integer getTotalOrderAmount();
+	Long getTotalOrderAmount();
 	
-	Integer getTotalMaterialCost();
+	Long getTotalMaterialCost();
 }
